requirements.txt

### 3. requirements.txt
Flask==2.3.2
Flask-Cors==3.0.10
mysql-connector-python==8.0.33
requests==2.31.0
wikipedia==1.4.0
python-dotenv==1.0.0
python-dateutil==2.9.0